@@ -1,4 +1,8 @@
-﻿namespace Sanchez.Processing.Extensions.Images;
+﻿using SixLabors.ImageSharp;
+using SixLabors.ImageSharp.PixelFormats;
+using SixLabors.ImageSharp.Processing;
+
+namespace Sanchez.Processing.Extensions.Images;
 
 public static class DrawImageExtensions
 {
