@@ -1,4 +1,7 @@
-﻿using SixLabors.ImageSharp.ColorSpaces.Conversion;
+﻿using SixLabors.ImageSharp;
+using SixLabors.ImageSharp.ColorSpaces.Conversion;
+using SixLabors.ImageSharp.PixelFormats;
+using SixLabors.ImageSharp.Processing;
 
 namespace Sanchez.Processing.ImageProcessing.Tint;
 
