@@ -9,21 +9,12 @@ public interface IImageMatcher
     List<Registration> FilterMatchingRegistrations(IEnumerable<Registration> registrations, DateTime targetTimestamp);
 }
 
-public class ImageMatcher : IImageMatcher
+public class ImageMatcher(ILogger<ImageMatcher> logger, RenderOptions options) : IImageMatcher
 {
-    private readonly ILogger<ImageMatcher> _logger;
-    private readonly RenderOptions _options;
-
-    public ImageMatcher(ILogger<ImageMatcher> logger, RenderOptions options)
-    {
-        _logger = logger;
-        _options = options;
-    }
-
     public List<Registration> FilterMatchingRegistrations(IEnumerable<Registration> registrations, DateTime targetTimestamp)
     {
-        _logger.LogInformation("Searching for images in {Path} with a timestamp of {Minutes} minutes tolerance to {Target}",
-            _options.SourcePath, _options.Tolerance.TotalMinutes, targetTimestamp);
+        logger.LogInformation("Searching for images in {Path} with a timestamp of {Minutes} minutes tolerance to {Target}",
+            options.SourcePath, options.Tolerance.TotalMinutes, targetTimestamp);
 
         // Return the closest match per satellite
         var matched = GetMatchedFiles(registrations, targetTimestamp);
@@ -39,7 +30,7 @@ public class ImageMatcher : IImageMatcher
 
     private IEnumerable<SatelliteMatch> GetMatchedFiles(IEnumerable<Registration> registrations, DateTime targetTimestamp)
     {
-        var tolerance = _options.Tolerance;
+        var tolerance = options.Tolerance;
 
         // Add files that have a timestamp within the tolerance of the selected time
         var matches =
@@ -51,15 +42,9 @@ public class ImageMatcher : IImageMatcher
         return matches;
     }
 
-    private sealed class SatelliteMatch
+    private sealed class SatelliteMatch(Registration registration, double deviation)
     {
-        public SatelliteMatch(Registration registration, double deviation)
-        {
-            Registration = registration;
-            Deviation = deviation;
-        }
-
-        internal Registration Registration { get; }
-        internal double Deviation { get; }
+        internal Registration Registration { get; } = registration;
+        internal double Deviation { get; } = deviation;
     }
 }
