@@ -4,17 +4,14 @@ using TimeZoneConverter;
 namespace Sanchez.Processing.Services.Filesystem.Parsers;
 
 /// <summary>
-///     Parses filenames of satellite imagery received from Electro-L no. 2. Note that Electro timestamps
-///     are in Russian Standard Time, not UTC.
+///     Parses filenames of satellite imagery received from Electro-L no. 2 and Electro-L no. 3. 
 /// </summary>
-public class ElectroFilenameParser : AbstractFilenameParser
+/// <remarks>
+///     Electro timestamps are in Russian Standard Time, not UTC.
+/// </remarks>
+public class ElectroFilenameParser(string? prefix, string? suffix) : AbstractFilenameParser
 {
-    protected override Regex Regex { get; }
+    protected override Regex Regex { get; } = new(prefix + "([0-9]{6}_[0-9]{4})_" + suffix + "\\.[^ ]*", RegexOptions.Compiled);
     protected override string TimestampFormat => "yyMMdd_HHmm";
     protected override TimeZoneInfo TimeZone => TZConvert.GetTimeZoneInfo("Russian Standard Time");
-
-    public ElectroFilenameParser(string? prefix, string? suffix) 
-    {
-        Regex = new Regex(prefix + "([0-9]{6}_[0-9]{4})_" + suffix + "\\.[^ ]*", RegexOptions.Compiled);
-    }
 }
