@@ -5,14 +5,9 @@ namespace Sanchez.Processing.Services.Filesystem.Parsers;
 /// <summary>
 ///     Parses filenames of satellite imagery received from GK-2A and processed by <c>xrit-rc</c>.
 /// </summary>
-public class Gk2AFilenameParser : AbstractFilenameParser
+public class Gk2AFilenameParser(string? prefix, string? suffix) : AbstractFilenameParser
 {
-    protected override Regex Regex { get; }
+    protected override Regex Regex { get; } = new(prefix + "_([0-9]{8}_[0-9]{6})" + suffix + "\\.[^ ]*", RegexOptions.Compiled);
 
     protected override string TimestampFormat => "yyyyMMdd_HHmmss";
-
-    public Gk2AFilenameParser(string? prefix, string? suffix)
-    {
-        Regex = new Regex(prefix + "_([0-9]{8}_[0-9]{6})" + suffix + "\\.[^ ]*", RegexOptions.Compiled);
-    }
 }
