@@ -5,14 +5,9 @@ namespace Sanchez.Processing.Services.Filesystem.Parsers;
 /// <summary>
 ///     Parses filenames of satellite imagery received from GOES-R satellites and processed by <c>goesproc</c>.
 /// </summary>
-public class GoesFilenameParser : AbstractFilenameParser
+public class GoesFilenameParser(string? prefix, string? suffix) : AbstractFilenameParser
 {
-    protected override Regex Regex { get; }
+    protected override Regex Regex { get; } = new(prefix + "([0-9]{8}T[0-9]{6}Z)" + suffix + "\\.[^ ]*", RegexOptions.Compiled);
 
     protected override string TimestampFormat => "yyyyMMddTHHmmssZ";
-
-    public GoesFilenameParser(string? prefix, string? suffix)
-    {
-        Regex = new Regex(prefix + "([0-9]{8}T[0-9]{6}Z)" + suffix + "\\.[^ ]*", RegexOptions.Compiled);
-    }
 }
