@@ -9,16 +9,12 @@ public interface ISatelliteImageLoader
     Activity RegisterImages(List<Registration> sourceRegistrations, DateTime? timestamp = null);
 }
 
-public class SatelliteImageLoader : ISatelliteImageLoader
+public class SatelliteImageLoader(IImageMatcher imageMatcher) : ISatelliteImageLoader
 {
-    private readonly IImageMatcher _imageMatcher;
-
-    public SatelliteImageLoader(IImageMatcher imageMatcher) => _imageMatcher = imageMatcher;
-
     public Activity RegisterImages(List<Registration> sourceRegistrations, DateTime? timestamp = null)
     {
         // If we are combining files by timestamp, locate all matching files based on the timestamp and tolerance
-        var registrations = timestamp == null ? sourceRegistrations : _imageMatcher.FilterMatchingRegistrations(sourceRegistrations, timestamp.Value);
+        var registrations = timestamp == null ? sourceRegistrations : imageMatcher.FilterMatchingRegistrations(sourceRegistrations, timestamp.Value);
         return new Activity(registrations);
     }
 }
