@@ -9,17 +9,8 @@ using WorkflowCore.Models;
 
 namespace Sanchez.Workflow.Steps.Equirectangular;
 
-internal sealed class GetVisibleRange : StepBody, IActivityStepBody
+internal sealed class GetVisibleRange(IProjectionOverlapCalculator projectionOverlapCalculator, RenderOptions options) : StepBody, IActivityStepBody
 {
-    private readonly IProjectionOverlapCalculator _projectionOverlapCalculator;
-    private readonly RenderOptions _options;
-
-    public GetVisibleRange(IProjectionOverlapCalculator projectionOverlapCalculator, RenderOptions options)
-    {
-        _projectionOverlapCalculator = projectionOverlapCalculator;
-        _options = options;
-    }
-
     public Activity? Activity { get; set; }
 
     public override ExecutionResult Run(IStepExecutionContext context)
@@ -27,15 +18,15 @@ internal sealed class GetVisibleRange : StepBody, IActivityStepBody
         ArgumentNullException.ThrowIfNull(Activity);
 
         // Calculates overlapping regions between satellites, or visible region if not stitching images.
-        _projectionOverlapCalculator.Initialise(Activity.Registrations.Select(p => p.Definition));
+        projectionOverlapCalculator.Initialise(Activity.Registrations.Select(p => p.Definition));
 
         // Set latitude and longitude ranges based on overlapping satellite ranges
         Activity.Registrations.ForEach(registration =>
         {
             registration.LatitudeRange = new ProjectionRange(registration.Definition.LatitudeRange);
             registration.LongitudeRange =
-                _options.StitchImages
-                    ? _projectionOverlapCalculator.GetNonOverlappingRange(registration.Definition)
+                options.StitchImages
+                    ? projectionOverlapCalculator.GetNonOverlappingRange(registration.Definition)
                     : new ProjectionRange(registration.Definition.LongitudeRange.UnwrapLongitude());
         });
 
