@@ -8,15 +8,11 @@ namespace Sanchez.Services;
 
 /// <summary>
 ///     Wraps log entries resulting from <see cref="ValidationException"/>, reporting
-///     as warning rather than error. This ensure that spurious Sentry.io reports aren't
-///     made for user error.
+///     as warning rather than error. This ensures that spurious Sentry.io reports aren't
+///     created for user errors.
 /// </summary>
-internal class ValidationExceptionWrapper : ILogEventSink, IDisposable
+internal class ValidationExceptionWrapper(ILogEventSink wrappedSink) : ILogEventSink, IDisposable
 {
-    private readonly ILogEventSink _wrappedSink;
-
-    public ValidationExceptionWrapper(ILogEventSink wrappedSink) => _wrappedSink = wrappedSink;
-
     public void Emit(LogEvent logEvent)
     {
         // Report validation exceptions are warnings
@@ -30,17 +26,17 @@ internal class ValidationExceptionWrapper : ILogEventSink, IDisposable
                 logEvent.Properties
                     .Select(kvp => new LogEventProperty(kvp.Key, kvp.Value)));
 
-            _wrappedSink.Emit(boosted);
+            wrappedSink.Emit(boosted);
         }
         else
         {
-            _wrappedSink.Emit(logEvent);
+            wrappedSink.Emit(logEvent);
         }
     }
 
     public void Dispose()
     {
-        (_wrappedSink as IDisposable)?.Dispose();
+        (wrappedSink as IDisposable)?.Dispose();
     }
 }
 
@@ -48,8 +44,8 @@ internal static class LoggerSinkConfigurationExtensions
 {
     /// <summary>
     ///     Wraps log entries resulting from <see cref="ValidationException"/>, reporting
-    ///     as warning rather than error. This ensure that spurious Sentry.io reports aren't
-    ///     made for user error.
+    ///     as warning rather than error. This ensures that spurious Sentry.io reports aren't
+    ///     created for user errors.
     /// </summary>
     internal static LoggerConfiguration ValidationWrapper(
         this LoggerSinkConfiguration lsc,
