@@ -37,7 +37,7 @@ public class GradientServiceTests : AbstractTests
         await CreateGradientFileAsync(gradient);
 
         var exception = Assert.Throws<ValidationException>(() => Service.GetGradient());
-        Assert.NotNull(exception);
+        exception.Should().NotBeNull();
         exception!.Message.Should().Be("Colour must be specified");
     }
 
@@ -60,7 +60,7 @@ public class GradientServiceTests : AbstractTests
         await CreateGradientFileAsync(gradient);
             
         var exception = Assert.Throws<ValidationException>(() => Service.GetGradient());
-        Assert.NotNull(exception);
+        exception.Should().NotBeNull();
         exception!.Message.Should().Be("Position must be specified for all colour stops");
     }
         
@@ -84,7 +84,7 @@ public class GradientServiceTests : AbstractTests
         await CreateGradientFileAsync(gradient);
                     
         var exception = Assert.Throws<ValidationException>(() => Service.GetGradient());
-        Assert.NotNull(exception);
+        exception.Should().NotBeNull();
         exception!.Message.Should().Be("-0.5 is an invalid position; valid values are from 0.0 to 1.0");
     }
 
@@ -108,7 +108,7 @@ public class GradientServiceTests : AbstractTests
         await CreateGradientFileAsync(gradient);
             
         var exception = Assert.Throws<ValidationException>(() => Service.GetGradient());
-        Assert.NotNull(exception);
+        exception.Should().NotBeNull();
         exception!.Message.Should().Be("2.3 is an invalid position; valid values are from 0.0 to 1.0");
     }
 
@@ -132,7 +132,7 @@ public class GradientServiceTests : AbstractTests
         await CreateGradientFileAsync(gradient);
             
         var exception = Assert.Throws<ValidationException>(() => Service.GetGradient());
-        Assert.NotNull(exception);
+        exception.Should().NotBeNull();
         exception!.Message.Should().Be("Unable to parse #sanchez as a hex triplet");
     }
 
